Added platform detection and Linux support.
With Homebrew's Linux support recently becoming official, the formula
needed to be updated to detect the platform and download the appropriate
binaries. This commit adds a simple platform detection mechanism using
Homebrew's OS module, which is undocumented but used by a large number
of core formulae to perform platform detection. This platform detection
is used to set the appropriate formula properties (in particular url and
sha256). More nuanced detection (e.g. of architecture) may be necessary
in the future as Homebrew expands its platform support, but this commit
lays the groundwork.

Fixes #1.