Updated formula TODO comments.
Our build script has come a long way since the deprecation of GOPATH and
the adoption of Go modules, leaving the TODO comment at the top of the
formula slightly out-of-date. This commit updates that TODO comment to
be more precise and removes the TODO from the older formulae since we
won't be updating them.

Signed-off-by: Jacob Howard <jacob@havoc.io>