Split devel block into separate formula.
Homebrew has officially deprecated devel blocks and now prints a warning
when dealing with formulae that define devel blocks. To work around
this, this commit splits the formula into nominal and "edge" versions.
The edge version will take the role of the devel version, but it's
actually a better organizational scheme since it's more explicit and
allows us to define additional release channels (e.g. a beta channel).

Signed-off-by: Jacob Howard <jacob@havoc.io>